port integration tests to jest with mocked webgl
