--- test/integration/index.html
@@ -1,36 +1,0 @@
-<!--
-    See https://mochajs.org/#running-mocha-in-the-browser for
-    a detailed explanation of the following page setup.
--->
-
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <title>VideoContext Integration Tests</title>
-        <link rel="stylesheet" media="all" href="../../node_modules/mocha/mocha.css">
-    </head>
-
-    <body>
-        <div id="mocha"></div>
-        <canvas id="videocontext-canvas" style="display:none;"></canvas>
-
-        <script src="../../node_modules/mocha/mocha.js"></script>
-        <script src="../../node_modules/chai/chai.js"></script>
-
-        <script>mocha.setup('bdd');</script>
-
-        <script src="../../dist/videocontext.js"></script>
-
-        <!-- Include all integration tests as separate `script` tags -->
-        <script src="videocontext.spec.js"></script>
-        <script src="sourcenode.spec.js"></script>
-
-        <script>
-            mocha.checkLeaks();
-            // Mocha errors on global variables by default
-            mocha.globals(['__VIDEOCONTEXT_REFS__']);
-            mocha.run();
-        </script>
-    </body>
-</html>
