--- test/integration/videocontext.spec.js
@@ -1,101 +1,0 @@
-/* eslint-disable */
-
-/**
- * These tests are run in the browser (and are written in ES5 for maximum support).
- * Mocha & Chai are the tools of choice here, since Jest (our unit-testing framework)
- * does not support brower-based tests.
- *
- * To run the tests, use `npm run test`. This will execute the `./test.sh` script
- * which runs the complete test suite, including unit and regression tests.
- */
-
-describe("VideoContext", function() {
-    describe("#effect()", function() {
-        var videocontext;
-
-        beforeEach(function() {
-            var canvas = document.getElementById("videocontext-canvas");
-            videocontext = new VideoContext(canvas);
-        });
-
-        it("should create an EffectNode from the passed definition", function() {
-            var effectNode = videocontext.effect(VideoContext.DEFINITIONS.MONOCHROME);
-
-            chai.assert.notEqual(effectNode.maximumConnections, Infinity);
-            chai.assert.deepEqual(effectNode.inputNames, ["u_image"]);
-        });
-    });
-
-    describe("#transition()", function() {
-        var videocontext;
-
-        beforeEach(function() {
-            var canvas = document.getElementById("videocontext-canvas");
-            videocontext = new VideoContext(canvas);
-        });
-
-        it("should create a TransitionNode from the passed definition", function() {
-            var transitionNode = videocontext.transition(VideoContext.DEFINITIONS.CROSSFADE);
-
-            chai.assert.notEqual(transitionNode.maximumConnections, Infinity);
-            chai.assert.notEqual(transitionNode.transition, undefined);
-            chai.assert.notEqual(transitionNode.clearTransitions, undefined);
-            chai.assert.deepEqual(transitionNode.inputNames, ["u_image_a", "u_image_b"]);
-        });
-    });
-
-    describe("#compositor()", function() {
-        var videocontext;
-
-        beforeEach(function() {
-            var canvas = document.getElementById("videocontext-canvas");
-            videocontext = new VideoContext(canvas);
-        });
-
-        it("should create a CompositingNode from the passed definition", function() {
-            var compositingNode = videocontext.compositor(VideoContext.DEFINITIONS.MONOCHROME);
-
-            chai.assert.equal(compositingNode.maximumConnections, Infinity);
-            chai.assert.deepEqual(compositingNode.inputNames, ["u_image"]);
-        });
-    });
-
-    describe("#duration", function() {
-        var videocontext;
-
-        beforeEach(function() {
-            var canvas = document.getElementById("videocontext-canvas");
-            videocontext = new VideoContext(canvas);
-        });
-
-        it("should return the time in seconds between time=0 and the stop time of the last SourceNode", function() {
-            var videoElement = document.createElement("video");
-            var videoNode1 = videocontext.video(videoElement);
-            var videoNode2 = videocontext.video(videoElement);
-
-            videoNode1.start(10);
-            videoNode1.stop(20.245);
-            videoNode2.start(0);
-            videoNode2.stop(10);
-            chai.assert.equal(videocontext.duration, 20.245);
-        });
-
-        it("should return Infinity if no stop time has been specified on one of the Nodes", function() {
-            var imageElement = document.createElement("img");
-            var imageNode = videocontext.createImageSourceNode(imageElement);
-
-            imageNode.start(0);
-            chai.assert.equal(videocontext.duration, Infinity);
-        });
-
-        it("should return 0 if all source nodes have had clearTimelineState called on them", function() {
-            var videoElement = document.createElement("video");
-            var videoNode = videocontext.video(videoElement);
-
-            videoNode.start(0);
-            videoNode.stop(10);
-            videoNode.clearTimelineState();
-            chai.assert.equal(videocontext.duration, 0);
-        });
-    });
-});
